plugin, config: add optional `maxmind_license_key` setting
As we have done with other plugins that ship with a built-in API key of
some kind (e.g. sopel-reddit), it's always preferable to include the
ability to override that built-in key before it's actually needed, lest
a key revocation break all installed plugin copies until an emergency
release can be published with configuration support.