plugin, config: MaxMind license key is now entirely optional
If no MaxMind key is provided, the plugin will download from a mirror of
the .mmdb files on GitHub.

With a license key, the plugin downloads directly from MaxMind.

I've also added error handling that didn't exist before, and though it's
not perfect, it's at least better. (There is now at least one case where
the function to find the GeoIP files can actually return False, which is
supposed to indicate an error to the call site inside the command func
but it was not actually possible to trigger that condition before.)

Note that I added the error handling because when I wrote this, the
project's MaxMind account was rate-limited for the day after some joker
on AWS used our whole download quota in a few minutes. Guess why I chose
to implement this change tonight?